feat(text-response): bug fixes
- render cells with formula differently and do not randomize by default
- clarify random number/date values are rounded down
- misc form fixes and refactoring
- fix RAND() in sheet function behavior
- remove dead code in client/app/lib/types.js
