+++ client/app/lib/components/form/fields/SingleFileInput/types.ts
@@ -1,0 +1,13 @@
+export type SpreadsheetCellValue = string | number | Date | undefined;
+
+export interface SheetGrid {
+  headers: string[];
+  rows: SpreadsheetCellValue[][];
+  rowOffset: number;
+  colOffset: number;
+}
+
+export interface GridData {
+  sheetNames: string[];
+  sheets: SheetGrid[];
+}
