+++ client/locales/ko.json
@@ -2418,7 +2418,7 @@
     "defaultMessage": "무작위 날짜 값 사용"
   },
   "course.assessment.question.textResponses.roundToDay": {
-    "defaultMessage": "일 단위로 반올림"
+    "defaultMessage": "일 단위로 내림 (00:00 UTC)"
   },
   "course.assessment.question.textResponses.shuffleRandomizationMode": {
     "defaultMessage": "다른 셀과 값 섞기"
